% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{Shiny UI for Urban Foraging}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\description{
Defines the UI for the Urban Foraging app.
}
\keyword{datasets}
